Refactor: better code-splitting
