+++ src/pages/UserInfo/index.tsx
@@ -1,20 +1,15 @@
-/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-
-/*
- * Due to this known issue: https://github.com/smooth-code/loadable-components/issues/173
- * Use .js extension for code-splitting file
- */
-
 import loadable from "@loadable/component";
 
 import { Loading, ErrorBoundary } from "../../components";
+import { Props, loadData } from "./UserInfo";
 
 const UserInfo = loadable(() => import("./UserInfo"), {
   fallback: <Loading />,
 });
 
-export default (props) => (
+export default (props: Props): JSX.Element => (
   <ErrorBoundary>
     <UserInfo {...props} />
   </ErrorBoundary>
 );
+export { loadData };
