--- src/pages/index.ts
@@ -1,7 +1,0 @@
-// @ts-expect-error
-import asyncHome from "./Home";
-// @ts-expect-error
-import asyncUserInfo from "./UserInfo";
-import NotFound from "./NotFound";
-
-export { asyncHome, asyncUserInfo, NotFound };
